Delete global_common.py, move its functions to dzutil
